move config.go -> config/toml.go
